+++ sources/template/Default/modules/modules_checkout_confirmation/cc_checkout_confirmation_delivery_address.php
@@ -28,7 +28,7 @@
       $this->description = CLICSHOPPING::getDef('module_checkout_confirmation_delivery_address_description');
 
       if (\defined('MODULE_CHECKOUT_CONFIRMATION_DELIVERY_ADDRESS_STATUS')) {
-        $this->sort_order = MODULE_CHECKOUT_CONFIRMATION_DELIVERY_ADDRESS_SORT_ORDER;
+        $this->sort_order = (int)MODULE_CHECKOUT_CONFIRMATION_DELIVERY_ADDRESS_SORT_ORDER;
         $this->enabled = (MODULE_CHECKOUT_CONFIRMATION_DELIVERY_ADDRESS_STATUS == 'True');
       }
      }
