+++ sources/template/Default/modules/modules_checkout_confirmation/cc_checkout_confirmation_law_hamon.php
@@ -29,7 +29,7 @@
       $this->description = CLICSHOPPING::getDef('module_checkout_confirmation_law_hamon_description');
 
       if (\defined('MODULE_CHECKOUT_CONFIRMATION_LAW_HAMON_STATUS')) {
-        $this->sort_order = MODULE_CHECKOUT_CONFIRMATION_LAW_HAMON_SORT_ORDER;
+        $this->sort_order = (int)MODULE_CHECKOUT_CONFIRMATION_LAW_HAMON_SORT_ORDER;
         $this->enabled = (MODULE_CHECKOUT_CONFIRMATION_LAW_HAMON_STATUS == 'True');
       }
      }
