+++ sources/template/Default/modules/modules_checkout_confirmation/cc_checkout_confirmation_payment_information.php
@@ -30,7 +30,7 @@
       $this->description = CLICSHOPPING::getDef('module_checkout_confirmation_payment_information_description');
 
       if (\defined('MODULE_CHECKOUT_CONFIRMATION_PAYMENT_INFORMATION_STATUS')) {
-        $this->sort_order = MODULE_CHECKOUT_CONFIRMATION_PAYMENT_INFORMATION_SORT_ORDER;
+        $this->sort_order = (int)MODULE_CHECKOUT_CONFIRMATION_PAYMENT_INFORMATION_SORT_ORDER;
         $this->enabled = (MODULE_CHECKOUT_CONFIRMATION_PAYMENT_INFORMATION_STATUS == 'True');
       }
      }
