+++ sources/template/Default/modules/modules_checkout_shipping/cs_checkout_shipping_button_process.php
@@ -29,7 +29,7 @@
       $this->description = CLICSHOPPING::getDef('module_checkout_shipping_button_process_description');
 
       if (\defined('MODULE_CHECKOUT_SHIPPING_BUTTON_PROCESS_STATUS')) {
-        $this->sort_order = MODULE_CHECKOUT_SHIPPING_BUTTON_PROCESS_SORT_ORDER;
+        $this->sort_order = (int)MODULE_CHECKOUT_SHIPPING_BUTTON_PROCESS_SORT_ORDER;
         $this->enabled = (MODULE_CHECKOUT_SHIPPING_BUTTON_PROCESS_STATUS == 'True');
       }
      }
