+++ sources/template/Default/modules/modules_checkout_success/chs_downloads.php
@@ -28,7 +28,7 @@
       $this->description = CLICSHOPPING::getDef('module_checkout_success_downloads_description');
 
       if (\defined('MODULE_CHECKOUT_SUCCESS_DOWNLOADS_STATUS')) {
-        $this->sort_order = \defined('MODULE_CHECKOUT_SUCCESS_DOWNLOADS_SORT_ORDER') ? MODULE_CHECKOUT_SUCCESS_DOWNLOADS_SORT_ORDER : 0;
+        $this->sort_order = \defined('MODULE_CHECKOUT_SUCCESS_DOWNLOADS_SORT_ORDER') ? (int)MODULE_CHECKOUT_SUCCESS_DOWNLOADS_SORT_ORDER : 0;
         $this->enabled = (MODULE_CHECKOUT_SUCCESS_DOWNLOADS_STATUS == 'True');
       }
     }
