+++ sources/template/Default/modules/modules_checkout_success/chs_redirect_old_order.php
@@ -27,7 +27,7 @@
       $this->description = CLICSHOPPING::getDef('module_checkout_success_redirect_old_order_description');
 
       if (\defined('MODULE_CHECKOUT_SUCCESS_REDIRECT_OLD_ORDER_STATUS')) {
-        $this->sort_order = \defined('MODULE_CHECKOUT_SUCCESS_REDIRECT_OLD_ORDER_SORT_ORDER') ? MODULE_CHECKOUT_SUCCESS_REDIRECT_OLD_ORDER_SORT_ORDER : 0;
+        $this->sort_order = \defined('MODULE_CHECKOUT_SUCCESS_REDIRECT_OLD_ORDER_SORT_ORDER') ? (int)MODULE_CHECKOUT_SUCCESS_REDIRECT_OLD_ORDER_SORT_ORDER : 0;
         $this->enabled = (MODULE_CHECKOUT_SUCCESS_REDIRECT_OLD_ORDER_STATUS == 'True');
       }
     }
