+++ sources/template/Default/modules/modules_checkout_success/chs_thank_you.php
@@ -27,7 +27,7 @@
       $this->description = CLICSHOPPING::getDef('module_checkout_success_thank_you_description');
 
       if (\defined('MODULE_CHECKOUT_SUCCESS_THANK_YOU_STATUS')) {
-        $this->sort_order = \defined('MODULE_CHECKOUT_SUCCESS_THANK_YOU_SORT_ORDER') ? MODULE_CHECKOUT_SUCCESS_THANK_YOU_SORT_ORDER : 0;
+        $this->sort_order = \defined('MODULE_CHECKOUT_SUCCESS_THANK_YOU_SORT_ORDER') ? (int)MODULE_CHECKOUT_SUCCESS_THANK_YOU_SORT_ORDER : 0;
         $this->enabled = (MODULE_CHECKOUT_SUCCESS_THANK_YOU_STATUS == 'True');
       }
     }
