+++ sources/template/Default/modules/modules_create_account_pro/cap_create_account_pro_registration.php
@@ -30,7 +30,7 @@
       $this->description = CLICSHOPPING::getDef('module_create_account_pro_registration_description');
 
       if (\defined('MODULE_CREATE_ACCOUNT_PRO_REGISTRATION_STATUS')) {
-        $this->sort_order = MODULE_CREATE_ACCOUNT_PRO_REGISTRATION_SORT_ORDER;
+        $this->sort_order = (int)MODULE_CREATE_ACCOUNT_PRO_REGISTRATION_SORT_ORDER;
         $this->enabled = (MODULE_CREATE_ACCOUNT_PRO_REGISTRATION_STATUS == 'True');
       }
     }
